requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
paddlepaddle==3.1.0
paddleocr==3.1.0
opencv-python==4.8.1.78
Pillow==10.1.0
pdfplumber==0.10.3
pymupdf==1.23.8
python-dateutil==2.8.2
regex==2023.10.3
ollama==0.5.1
docling